requirements.txt
google-genai>=1.2.0
apscheduler>=3.10.4
fastapi>=0.109.0
uvicorn>=0.27.0
pillow>=10.2.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0
httpx>=0.27.0
pydantic>=2.0.0
json5>=0.9.0
psycopg2-binary>=2.9.9
duckduckgo_search>=5.0.0
requests>=2.31.0
rembg>=2.0.56
mlflow>=2.21.0
pandas>=2.2.3
openpyxl>=3.1.5
